Fix component
1. Delete unnecessary Page component
2. Separate App component from index.jsx
