--- src/components/Page.jsx
@@ -1,15 +1,0 @@
-import Counter from './Counter';
-import Buttons from './Buttons';
-
-export default function Page({ count, onClick }) {
-  return (
-    <div>
-      <p>Counter</p>
-      <Counter
-        count={count}
-        onClick={onClick}
-      />
-      <Buttons onClick={onClick} />
-    </div>
-  );
-}
